Collapse C079 into C010
