+++ crates/shuck-linter/resources/test/fixtures/correctness/C010.sh
@@ -15,8 +15,15 @@ else
   printf '%s\n' miss
 fi
 
-# Valid: non-test fallthrough chains and assignment ternaries belong to other rules.
+# Invalid: general fallthrough chains are part of the same warning family.
+[ "$x" = foo ] && printf '%s\n' yes || rm -f no
 true && false || printf '%s\n' fallback
+
+# Valid: common status-propagation and formatter idioms stay exempt.
+cond && return 0 || return 1
+ready && printf '%s\n' on || printf '%s\n' off
+
+# Valid: assignment ternaries belong to other rules.
 [ -n "$x" ] && out=foo || out=bar
 
 # Valid: pure && and pure || chains are not this rule.
