+++ crates/shuck-linter/src/checker.rs
@@ -660,9 +660,6 @@ impl<'a> Checker<'a> {
         if self.is_rule_enabled(Rule::ChainedTestBranches) {
             rules::correctness::chained_test_branches::chained_test_branches(self);
         }
-        if self.is_rule_enabled(Rule::ShortCircuitFallthrough) {
-            rules::correctness::short_circuit_fallthrough::short_circuit_fallthrough(self);
-        }
         if self.is_rule_enabled(Rule::DefaultElseInShortCircuit) {
             rules::correctness::default_else_in_short_circuit::default_else_in_short_circuit(self);
         }
