+++ crates/shuck-linter/src/registry.rs
@@ -170,12 +170,6 @@ declare_rules! {
         RedirectClobbersInput
     ),
     (
-        "C079",
-        Category::Correctness,
-        Severity::Warning,
-        ShortCircuitFallthrough
-    ),
-    (
         "C082",
         Category::Correctness,
         Severity::Warning,
@@ -820,7 +814,8 @@ pub fn code_to_rule(code: &str) -> Option<Rule> {
         "SH-047" => Some(Rule::SshLocalExpansion),
         "SH-151" => Some(Rule::EvalOnArray),
         "SH-045" => Some(Rule::ChainedTestBranches),
-        "SH-201" => Some(Rule::ShortCircuitFallthrough),
+        "C079" => Some(Rule::ChainedTestBranches),
+        "SH-201" => Some(Rule::ChainedTestBranches),
         "SH-046" => Some(Rule::LineOrientedInput),
         "SH-048" => Some(Rule::LeadingGlobArgument),
         "SH-049" => Some(Rule::FindOutputLoop),
@@ -1163,7 +1158,8 @@ mod tests {
         assert_eq!(code_to_rule("SH-044"), Some(Rule::RmGlobOnVariablePath));
         assert_eq!(code_to_rule("SH-047"), Some(Rule::SshLocalExpansion));
         assert_eq!(code_to_rule("SH-045"), Some(Rule::ChainedTestBranches));
-        assert_eq!(code_to_rule("SH-201"), Some(Rule::ShortCircuitFallthrough));
+        assert_eq!(code_to_rule("C079"), Some(Rule::ChainedTestBranches));
+        assert_eq!(code_to_rule("SH-201"), Some(Rule::ChainedTestBranches));
         assert_eq!(code_to_rule("SH-046"), Some(Rule::LineOrientedInput));
         assert_eq!(code_to_rule("SH-049"), Some(Rule::FindOutputLoop));
         assert_eq!(code_to_rule("C113"), Some(Rule::FindOutputLoop));
