+++ crates/shuck-linter/src/rules/correctness/chained_test_branches.rs
@@ -1,5 +1,6 @@
-use crate::facts::MixedShortCircuitKind;
+use crate::facts::{ListFact, ListSegmentKind, MixedShortCircuitKind};
 use crate::{Checker, Rule, Violation};
+use shuck_ast::BinaryOp;
 
 pub struct ChainedTestBranches;
 
@@ -18,11 +19,122 @@ pub fn chained_test_branches(checker: &mut Checker) {
         .facts()
         .lists()
         .iter()
-        .filter(|list| list.mixed_short_circuit_kind() == Some(MixedShortCircuitKind::TestChain))
+        .filter(|list| matches_mixed_short_circuit(checker, list))
         .filter_map(|list| list.mixed_short_circuit_span())
         .collect::<Vec<_>>();
 
     checker.report_all_dedup(spans, || ChainedTestBranches);
+}
+
+fn matches_mixed_short_circuit(checker: &Checker<'_>, list: &ListFact<'_>) -> bool {
+    match list.mixed_short_circuit_kind() {
+        Some(MixedShortCircuitKind::TestChain) => true,
+        Some(MixedShortCircuitKind::Fallthrough) => !list_exempts_warning(checker, list),
+        _ => false,
+    }
+}
+
+fn list_exempts_warning(checker: &Checker<'_>, list: &ListFact<'_>) -> bool {
+    if matches_status_propagation_assignment(list) {
+        return true;
+    }
+
+    if matches_exempt_fallback_branch(checker, list) {
+        return true;
+    }
+
+    let Some(branch_names) = ternary_branch_names(checker, list) else {
+        return false;
+    };
+
+    if branch_names
+        .iter()
+        .all(|name| matches!(name, Some("return" | "exit")))
+    {
+        return true;
+    }
+
+    matches!(
+        &branch_names,
+        [Some("echo"), Some("echo")] | [Some("printf"), Some("printf")]
+    )
+}
+
+fn matches_status_propagation_assignment(list: &ListFact<'_>) -> bool {
+    if !matches_and_or_ternary(list) {
+        return false;
+    }
+
+    let segments = list.segments();
+    let [first, _, last] = segments else {
+        return false;
+    };
+
+    first.kind() == ListSegmentKind::AssignmentOnly
+        && last.kind() == ListSegmentKind::AssignmentOnly
+        && first.assignment_target().is_some()
+        && first.assignment_target() == last.assignment_target()
+}
+
+fn matches_exempt_fallback_branch(checker: &Checker<'_>, list: &ListFact<'_>) -> bool {
+    let Some(last_operator) = list.operators().last() else {
+        return false;
+    };
+    if last_operator.op() != BinaryOp::Or {
+        return false;
+    }
+
+    let Some(last_segment) = list.segments().last() else {
+        return false;
+    };
+
+    if last_segment.kind() == ListSegmentKind::AssignmentOnly {
+        return true;
+    }
+
+    matches!(
+        checker
+            .facts()
+            .command(last_segment.command_id())
+            .effective_or_literal_name(),
+        Some("return" | "exit" | "true" | ":")
+    )
+}
+
+fn ternary_branch_names<'a>(
+    checker: &'a Checker<'a>,
+    list: &ListFact<'a>,
+) -> Option<[Option<&'a str>; 2]> {
+    if !matches_and_or_ternary(list) {
+        return None;
+    }
+
+    let [_, then_branch, else_branch] = list.segments() else {
+        return None;
+    };
+
+    Some([
+        checker
+            .facts()
+            .command(then_branch.command_id())
+            .effective_or_literal_name(),
+        checker
+            .facts()
+            .command(else_branch.command_id())
+            .effective_or_literal_name(),
+    ])
+}
+
+fn matches_and_or_ternary(list: &ListFact<'_>) -> bool {
+    list.segments().len() == 3
+        && matches!(
+            list.operators()
+                .iter()
+                .map(|operator| operator.op())
+                .collect::<Vec<_>>()
+                .as_slice(),
+            [BinaryOp::And, BinaryOp::Or]
+        )
 }
 
 #[cfg(test)]
@@ -36,6 +148,7 @@ mod tests {
 [ \"$x\" = foo ] && [ \"$x\" = bar ] || [ \"$x\" = baz ]
 false || true && [ \"$x\" = baz ]
 true && false; false || printf '%s\\n' ok
+[ \"$dir\" = vendor ] && mv go-* \"$dir\" || mv pkg-* \"$dir\"
 [ -n \"$x\" ] && out=foo || out=bar
 ";
         let diagnostics =
@@ -46,7 +159,76 @@ true && false; false || printf '%s\\n' ok
                 .iter()
                 .map(|diagnostic| diagnostic.span.slice(source))
                 .collect::<Vec<_>>(),
-            vec!["&&", "||"]
+            vec!["&&", "||", "&&"]
         );
+    }
+
+    #[test]
+    fn ignores_status_propagation_and_formatter_idioms() {
+        let source = "\
+cond && return 0 || return 1
+return_code=0 && cmd >out 2>err || return_code=$?
+is_tty && printf '%s\\n' a || printf '%s\\n' b
+flag && echo enabled || echo disabled
+";
+        let diagnostics =
+            test_snippet(source, &LinterSettings::for_rule(Rule::ChainedTestBranches));
+
+        assert!(diagnostics.is_empty());
+    }
+
+    #[test]
+    fn only_ignores_three_segment_status_propagation_shapes() {
+        let source = "\
+rc=0 || run && rc=$?
+rc=0 && run || fallback && rc=$?
+";
+        let diagnostics =
+            test_snippet(source, &LinterSettings::for_rule(Rule::ChainedTestBranches));
+
+        assert_eq!(
+            diagnostics
+                .iter()
+                .map(|diagnostic| diagnostic.span.slice(source))
+                .collect::<Vec<_>>(),
+            vec!["||", "&&"]
+        );
+    }
+
+    #[test]
+    fn ignores_final_assignment_and_status_fallback_idioms() {
+        let source = "\
+check_download_exists \"$path\" && download_status=0 || download_status=$?
+__rvm_select_set_variable_defaults && __rvm_select_after_parse || return $?
+test -d x && mv x y || :
+test -d x && mv x y || true
+";
+        let diagnostics =
+            test_snippet(source, &LinterSettings::for_rule(Rule::ChainedTestBranches));
+
+        assert!(diagnostics.is_empty());
+    }
+
+    #[test]
+    fn ignores_when_only_the_fallback_is_an_assignment() {
+        let source = "\
+[ -n \"$x\" ] && run_task || fallback=1
+";
+        let diagnostics =
+            test_snippet(source, &LinterSettings::for_rule(Rule::ChainedTestBranches));
+
+        assert!(diagnostics.is_empty());
+    }
+
+    #[test]
+    fn still_reports_when_the_fallback_is_not_an_exempt_control_flow_or_assignment() {
+        let source = "\
+[ -n \"$x\" ] && run_task || fallback
+";
+        let diagnostics =
+            test_snippet(source, &LinterSettings::for_rule(Rule::ChainedTestBranches));
+
+        assert_eq!(diagnostics.len(), 1);
+        assert_eq!(diagnostics[0].span.slice(source), "&&");
     }
 }
