+++ crates/shuck-linter/src/suppression/shellcheck_map.rs
@@ -312,7 +312,7 @@ impl Default for ShellCheckCodeMap {
             // ShellCheck 0.11.0 reports `find -exec` pre-expansion warnings as SC2014.
             // Keep SC2295 as a suppression alias for authored C078 metadata.
             (2014, Rule::UnquotedGlobsInFind),
-            (2296, Rule::ShortCircuitFallthrough),
+            (2296, Rule::ChainedTestBranches),
             // ShellCheck 0.11.0 reports loop-list glob+expansion mixes as SC2231.
             // Keep SC2349 as a suppression alias for authored C114 metadata.
             (2231, Rule::GlobWithExpansionInLoop),
@@ -732,7 +732,7 @@ impl Default for ShellCheckCodeMap {
                     (2013, Rule::LineOrientedInput),
                     (2015, Rule::ChainedTestBranches),
                     (2014, Rule::UnquotedGlobsInFind),
-                    (2296, Rule::ShortCircuitFallthrough),
+                    (2296, Rule::ChainedTestBranches),
                     (2231, Rule::GlobWithExpansionInLoop),
                     (2022, Rule::GlobInGrepPattern),
                     (2062, Rule::UnquotedGrepRegex),
@@ -946,11 +946,6 @@ impl Default for ShellCheckCodeMap {
                 (2372, Rule::SingleLetterCaseLabel),
                 (2382, Rule::GetoptsOptionNotInCase),
                 (2383, Rule::CaseArmNotInGetopts),
-                // The pinned ShellCheck oracle still reports ordinary `A && B || C`
-                // fallthrough chains as SC2015. Keep that older code as a
-                // compatibility alias so targeted large-corpus validation can
-                // compare C079 against the actual oracle output.
-                (2015, Rule::ShortCircuitFallthrough),
                 (2114, Rule::ConditionalAssignmentShortcut),
                 (2165, Rule::SingleIterationLoop),
                 (2322, Rule::SuWithoutFlag),
@@ -1375,11 +1370,8 @@ mod tests {
         assert_eq!(map.resolve("SC2326"), Some(Rule::GlobAssignedToVariable));
         assert_eq!(map.resolve_all("SC2014"), vec![Rule::UnquotedGlobsInFind]);
         assert_eq!(map.resolve_all("SC2295"), vec![Rule::UnquotedGlobsInFind]);
-        assert_eq!(
-            map.resolve_all("SC2015"),
-            vec![Rule::ChainedTestBranches, Rule::ShortCircuitFallthrough]
-        );
-        assert_eq!(map.resolve("SC2296"), Some(Rule::ShortCircuitFallthrough));
+        assert_eq!(map.resolve_all("SC2015"), vec![Rule::ChainedTestBranches]);
+        assert_eq!(map.resolve("SC2296"), Some(Rule::ChainedTestBranches));
         assert_eq!(
             map.resolve_all("SC2231"),
             vec![Rule::GlobWithExpansionInLoop]
@@ -1874,8 +1866,7 @@ mod tests {
             (2258, Rule::BareRead),
             (2013, Rule::LineOrientedInput),
             (2015, Rule::ChainedTestBranches),
-            (2015, Rule::ShortCircuitFallthrough),
-            (2296, Rule::ShortCircuitFallthrough),
+            (2296, Rule::ChainedTestBranches),
             (2016, Rule::SingleQuotedLiteral),
             (2014, Rule::UnquotedGlobsInFind),
             (2231, Rule::GlobWithExpansionInLoop),
