+++ docs/bugs/C010.md
@@ -1,47 +1,35 @@
-# C010: Re-scope large-corpus comparison for the SC2015 split
+# C010: Collapse mixed short-circuit parity onto one rule
 
 ## Rule
 
 | Field | Value |
 |-------|-------|
 | Shuck code | C010 |
-| ShellCheck code | SC2015 |
+| ShellCheck code | SC2015 primary, SC2296 alias |
 | Category | Correctness |
 | Rule file | `crates/shuck-linter/src/rules/correctness/chained_test_branches.rs` |
 
-## Delta Snapshot
+## Summary
 
-Re-sampled on 2026-04-16 using `make test-large-corpus SHUCK_LARGE_CORPUS_RULES=C010 SHUCK_LARGE_CORPUS_SAMPLE_PERCENT=10 SHUCK_LARGE_CORPUS_KEEP_GOING=1`.
+The previous split between `C010` and `C079` matched two internal fact kinds, but it left the large-corpus harness comparing one live ShellCheck warning family against two Shuck rules. The refactor removes that user-facing split.
 
-That run covered 682 fixtures, skipped 79 unsupported-shell fixtures, and reported only 2 blocking implementation diffs:
+`C010` now owns both:
 
-- `rvm__rvm__scripts__pkg:52`
-- `void-linux__void-packages__srcpkgs__inadyn__files__inadyn__run:5`
-
-## Current Analysis
-
-Both remaining `shellcheck-only` hits are ordinary fallthrough chains, not the all-condition shortcut shape that `C010` is written to flag.
-
-- `rvm__rvm__scripts__pkg:52` uses `__rvm_rm_rf ... && rvm_log ... || rvm_error ...`
-- `void-linux__void-packages__srcpkgs__inadyn__files__inadyn__run:5` uses a longer `test && mkdir && chown || rm` chain
-
-Those patterns already belong to `C079` (`ShortCircuitFallthrough`). The pinned ShellCheck build used by the large-corpus harness currently emits both the narrow `C010` family and the broader `C079` family as `SC2015`, so a selected `C010` corpus run cannot isolate the intended all-condition subset.
-
-This means the remaining failures are not evidence that `C010` itself should expand. Broadening `C010` here would collapse the authored split between:
-
-- `C010`: mixed `&&` / `||` chains where every branch is a condition-like check
-- `C079`: general fallthrough chains where later commands can fail and unexpectedly trigger the fallback
+- condition-only shortcut chains
+- broader `A && B || C` fallthrough chains after the existing exemptions
 
 ## Resolution
 
-Record a `comparison_target_notes` entry for `C010` so the large-corpus harness no longer treats the live `SC2015` bucket as a stable standalone comparison target for this rule. Keep the rule implementation unchanged and let `C079` own the broader live-oracle fallthrough parity in this ShellCheck build.
+Keep the internal fact classification, but report both shapes as `C010`. Accept both `SC2015` and `SC2296` as compatibility codes for the same rule and retire the separate `C079` rule surface.
+
+With the fallback exemptions aligned to the nix-pinned ShellCheck oracle, the targeted 10% corpus sample for `C010` now completes with zero reviewed divergences and no comparison metadata file.
 
 ## Checklist
 
-- [x] Confirm the remaining selected-run failures are broader fallthrough chains, not missed all-condition `C010` sites.
-- [x] Record the comparison-target note in `crates/shuck/tests/testdata/corpus-metadata/c010.yaml`.
-- [x] Re-run the 10% corpus sample and update this document.
+- [x] Fold fallthrough reporting into `C010`.
+- [x] Retire the separate `C079` rule entry and remap its compatibility aliases.
+- [x] Re-run the targeted and large-corpus parity checks after the refactor.
 
 ## Notes
 
-The existing broad `shuck-only` reviewed divergence in `c010.yaml` still documents the intentional stricter policy on helper-library and test-harness shortcut chains. This change only fixes the selected-rule comparison target so `C010` no longer gets judged against the broader live `SC2015` bucket that now also feeds `C079`.
+The internal facts still distinguish condition-only chains from fallthrough chains. The change is only to the user-facing and corpus-comparison surface.
