--- docs/rules/C079.yaml
@@ -1,23 +1,0 @@
-legacy_code: SH-201
-legacy_name: short-circuit-fallthrough
-new_category: Correctness
-new_code: C079
-runtime_kind: ast
-shellcheck_code: SC2296
-shells:
-  - sh
-  - bash
-  - dash
-  - ksh
-description: "A `&&`/`||` chain may not branch as intended."
-rationale: "Use an explicit `if`/`else` to avoid unexpected fallthrough."
-source:
-  shell_checks_rule: rules/SH-201.yaml
-  shell_checks_example: examples/201.sh
-examples:
-  - kind: invalid
-    source: shell-checks/examples/201.sh
-    code: |
-      #!/bin/bash
-      # shellcheck disable=2086
-      [ $DIR = vendor ] && mv go-* $DIR || mv pkg-* $DIR
